transformed acceleration measurement to m/s

+++ raisavis/src/main/java/raisa/vis/Vector3D.java
@@ -2,23 +2,23 @@ package raisa.vis;
 
 public class Vector3D {
 
-	private int x, y, z;
+	private float x, y, z;
 
-	public Vector3D(int x, int y, int z) {
+	public Vector3D(float x, float y, float z) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
 	}
 	
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 	
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 	
-	public int getZ() {
+	public float getZ() {
 		return z;
 	}
